backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.8.0
pydantic==2.9.2
pydantic-settings==2.1.0
pytest==8.3.5
pytest-asyncio==0.23.3
httpx==0.27.0
openai==1.57.4
cohere==5.15.0
psycopg[binary,pool]==3.2.4
python-multipart==0.0.21
python-dotenv==1.0.1
tiktoken==0.8.0
sqlalchemy==2.0.35
asyncpg==0.31.0